requirements.txt
actions-toolkit==0.1.15
attrs==22.2.0
autopep8==2.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pycparser==2.21
PyGithub==1.56
PyJWT==2.6.0
PyNaCl==1.5.0
pytest==7.2.1
pytest-dotenv==0.5.2
python-dotenv==0.21.1
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
wrapt==1.14.1
